added library Sigmajs containing folder GraphEditor
Sigmajs is a JavaScript library, see also sigma.js-1.2.1/LICENSE.txt .

It is used to develop a "GraphEditor" for creating Graphs after the users ideas. The goal is to import these graphs to the main program, to calculate IDEs and display them with the specified layout and data.

The current versions can be found in sigma.js-1.2.1/GraphEditor , but the work is still in progress and the editor is not yet working as intended.