added standard values for some variables and reworked impot of graphs via .gexf-files
The variable 'alpha' now gets initialized to 1/2 and the variable 'variante' to 'A', if they are not specified in 'Math\data.py'.

The variables 'kanten_queue', 'start_queue', 'ende_queue' and 'y0_queue' are now initialized to empty lists, if at least one of them is not specified in 'Math\data.py'.

If no graph is specified in 'Math\data.py', the program no reads the graph from 'GraphenGEXF\myGraph.gexf' correctly.