more bug fixes
resolved a bug, where the change of node labels was used for calculations before it was correctly calculated for the current time step, instead of the other way around. Also a truly active edge is now used to calculate the change of cost of an active edge, instead of an edge that was only active in the beginning of the phase.